+++ src/model/equations/Plane.cpp
@@ -1,9 +1,9 @@
-#include "../../../include/model/Equations/Plane.h"
+#include "Plane.h"
 
 
-equations::Plane::Plane(float A, float B, float C, float D) : A(A), B(B), C(C), D(D) {}
+equations::Plane::Plane(double A, double B, double C, double D) : A(A), B(B), C(C), D(D) {}
 
-equations::Plane::Plane(std::vector<float> in) : A(in[0]), B(in[1]), C(in[2]), D(in[3]) {}
+equations::Plane::Plane(std::vector<double> in) : A(in[0]), B(in[1]), C(in[2]), D(in[3]) {}
 
 equations::Plane::Plane(pcl::ModelCoefficients::Ptr in) : A(in->values[0]), B(in->values[1]),
                                                           C(in->values[2]), D(in->values[3]) {}
